refactor(region): use class for consistency, now also mark in the endregion
