+++ README.md
@@ -44,7 +44,6 @@ Check the full details in [logger docs](./src/docs/logger.md)
 
 6. ***Region***
 
-
 Automatic code region markers, and delete markers easily in your editor/workspace.
 
 Check the full details in [region docs](./src/docs/region.md)
